requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
alphabet_detector
apscheduler==3.6.3
asyncio==3.4.3
beautifulsoup4==4.12.3
bing_image_downloader==1.1.2
bleach==6.1.0
bs4==0.0.2
cachetools==4.2.2
cloudscraper==1.2.71
countryinfo==0.1.2
currencyconverter==0.17.23
dateparser==1.2.0
emoji==2.12.1
emoji-country-flag==1.3.2
faker
feedparser==6.0.11
ffmpeg-python==0.2.0
future==0.18.3
geopy==2.4.1
gtts
gpytranslate==1.5.1
hachoir==3.3.0
heroku3
html2text==2024.2.26
jikanpy==4.3.2
lxml==4.9.4
markdown2==2.4.13
MukeshAPI
motor==3.3.2
nekos.py
pydictionary==2.0.1
pyfiglet
pykeyboard==0.1.5
pymongo
pynewtonmath==1.0.1
pyrate-limiter==2.10.0
pyshorteners
pyrogram==2.0.106
python-telegram-bot==13.15
regex==2024.5.15
requests==2.32.3
search_engine_parser==0.6.8
secureme==0.0.4
speedtest-cli==2.1.3
sqlalchemy==1.4.46
telegraph==2.2.0
telethon==1.36.0
tgcrypto==1.2.5
wikipedia==1.4.0
wget
yt_dlp
youtube_search
youtube_search_python

